v 1.0.0

+++ poker/game.py
@@ -3,12 +3,12 @@ from poker.player import Player
 
 
 class Game:
-    def __init__(self, sb: float, bb: float):
+    def __init__(self, sb: int, bb: int):
         self.sb = sb
         self.bb = bb
         self.players = []
 
-    def add_player(self, name: str, stack: float, is_bot: bool = True, player_type: str = 'rand') -> None:
+    def add_player(self, name: str, stack: int, is_bot: bool = True, player_type: str = 'rand') -> None:
         self.players.append(Player(name=name, stack=stack, is_bot=is_bot, player_type=player_type))
 
     def print_players(self) -> None:
